pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ocr-data-preparation-pipeline"
version = "0.0.1"
authors = [
  { name="ta4tsering", email="tashitserin@dharmaduta.in" },
]
description = "OCR data preparation pipeline for processing images"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pydantic>=2.0.0",
    "Pillow>=10.0.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pre-commit",
]

[project.urls]
"Homepage" = "https://github.com/ta4tsering/OCR-data-preparation-pipeline"
"Bug Tracker" = "https://github.com/ta4tsering/OCR-data-preparation-pipeline/issues"
